// src/app/interfaces-enums/enums.ts
export enum Position{
    President = "President", 
    ViceP = "Vice-President",
    Secretary = "Secretary",
    Treasurer = "Treasurer",
    ConfChair = "Conference Planning Chair",
    FinChair = "Finance Chair",
    MemChair = "Membership Chair",
    PCIChair = "PCI Chair",
    Parl = "Parlimentarian",
    Programs = "Programs Chair",
    PubChair = "Public Relations Chair",
    Senator = "Senator",
  }

  export enum Classification{
    Freshman = "Freshman", 
    Sophmore = "Sophmore",
    Junior = "Junior",
    Senior = "Senior",
    fifth = "5th Year",
    other = "Other"
  }